+++ packages/compiler-cli/ngcc/src/host/delegating_host.ts
@@ -11,7 +11,7 @@ import * as ts from 'typescript';
 import {ClassDeclaration, ClassMember, CtorParameter, Declaration, Decorator, FunctionDefinition, Import, ReflectionHost} from '../../../src/ngtsc/reflection';
 import {isFromDtsFile} from '../../../src/ngtsc/util/src/typescript';
 
-import {ModuleWithProvidersFunction, NgccClassSymbol, NgccReflectionHost, SwitchableVariableDeclaration} from './ngcc_host';
+import {NgccClassSymbol, NgccReflectionHost, SwitchableVariableDeclaration} from './ngcc_host';
 
 /**
  * A reflection host implementation that delegates reflector queries depending on whether they
@@ -147,10 +147,6 @@ export class DelegatingReflectionHost implements NgccReflectionHost {
 
   getDecoratorsOfSymbol(symbol: NgccClassSymbol): Decorator[]|null {
     return this.ngccHost.getDecoratorsOfSymbol(symbol);
-  }
-
-  getModuleWithProvidersFunctions(sf: ts.SourceFile): ModuleWithProvidersFunction[] {
-    return this.ngccHost.getModuleWithProvidersFunctions(sf);
   }
 
   getSwitchableDeclarations(module: ts.Node): SwitchableVariableDeclaration[] {
